Applying Policies and Middleware

+++ app/Http/Controllers/NoticeController.php
@@ -13,6 +13,7 @@ class NoticeController extends Controller
     public function index(Course $course)
     {
 
+        $this->authorize('view', $course);
         $notices = Notice::where('course_id', $course->id)->get();
         return view('teacher.courses.notice', [
             'notices' => $notices,
@@ -22,11 +23,11 @@ class NoticeController extends Controller
 
     public function store(Course $course, Request $request)
     {
+        $this->authorize('view', $course);
         $request->validate([
             'subject' => 'required | min:3',
             'description' => 'required | min:3'
         ]);
-
         $notice = new Notice;
         $notice->subject = $request->subject;
         $notice->description = $request->description;
@@ -35,9 +36,9 @@ class NoticeController extends Controller
         session()->flash('notice_created', "Notice Created Successfully");
         return back();
     }
-
     public function destroy(Notice $notice)
     {
+        $this->authorize('delete', $notice);
         $notice->delete($notice->id);
         session()->flash('deleted', "Notice Deleted Successfully");
         return back();
@@ -45,20 +46,30 @@ class NoticeController extends Controller
 
     public function show(User $user)
     {
-
         $user->load(['enroll.payment' => function ($query) use ($user) {
             $query->where('status', Config('constants.payment.approved'))->with('course.notices');
         }]);
         return view('student.notice', ['data' => $user]);
 
+
+
+        // $notices = User::whereId($current_id)->with(['payments' => function ($query) use ($user) {
+        //     return $query->where('status', Config('constants.payment.approved'))->with('notices');
+        // }])->get();
+
+        // return $notices->payments;
+        // return view('student.notice', ['notices' => $notices]);
+
+        // return ($notices);
+
+
+      
+
+
         // $user->load(['enroll.payment' => function ($query) use ($user) {
         //     $query->where('status', Config('constants.payment.approved'));
         // }])->with('course.notices');
         // return view('student.notice', ['data' => $user]); // Working 
-
-
-
-
         // $notices = Notice::whereHas('payment', function ($query) use ($user) {
         //     $query->where('status', Config('constants.payment.approved'))
         //     ->where('student_id', $user->id);
@@ -66,8 +77,6 @@ class NoticeController extends Controller
         //     ->with(['payment.course'])
         //     ->get();
         // $filteredNotices = $notices->pluck('payment.course.notices')->flatten();
-
-
         // return $filteredNotices;
         //     return $user->enroll;
 
@@ -88,21 +97,13 @@ class NoticeController extends Controller
         // }
         // exit;
         // Enroll->payment->course->notice
-
-
         // $user->load(['enroll.notices.course.payment' => function ($query) use ($user) {
         //     $query->where('status', Config('constants.payment.approved'));
         // }]);
-
-
         // return view('student.notice', ['user' => $user]);
-
-
         // $user->load(['enroll.notices.course.payment' => function ($query) {
         //     $query->where('status', Config('constants.payment.approved'));
         // }]);
-
-
         // $user->payments->course->load(['notices']);
         // return $user;
         // $payments =  $user->payments()->where('status', Config('constants.payment.approved'))->get();
@@ -120,4 +121,19 @@ class NoticeController extends Controller
 
 
     }
+
+    // public function requireee(Request $request)
+    // {
+    //     $rules = [
+    //         'name' => ['required', 'min:3'],
+    //     ];
+
+    //     if ($request->input('name') === 'keshav') {
+    //         $rules['state'][] = 'required';
+    //     }
+    //     $validatedData = $request->validate($rules);
+    //     return $validatedData;
+    // }
+
+
 }
